Improve order validation and close #30
